refactor: separate repository package
