+++ src/main/kotlin/com/freewill/repository/jpa/RegionRepository.kt
@@ -1,4 +1,4 @@
-package com.freewill.repository
+package com.freewill.repository.jpa
 
 import com.freewill.entity.Region
 import org.springframework.data.jpa.repository.JpaRepository
